+++ src/config.ts
@@ -6,7 +6,8 @@ export { IDL } from "./types/zo";
 export const CONTROL_ACCOUNT_SIZE = 8 + 4482;
 
 // == PARAMS ==
-export const ZO_TAKER_FEE = 0.001; // 10bps
+export const ZO_FUTURE_TAKER_FEE = 0.001; // 10bps
+export const ZO_OPTION_TAKER_FEE = 0.001; // temporary value, not actual
 export const USDC_DECIMALS = 6;
 
 // == 01 PROGRAM IDS ==