+++ CHANGELOG.md
@@ -11,7 +11,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Withdraw & deposit now accept mint and deposit/withdraw directly to/from associated token account
 - Associated token account is created before withdrawal if such account does not exist
 - Old implementations of deposit / withdrawals are repurposed as depositRaw/ withdrawRaw in case users want to deposit/withdraw to/from a specific account
-- Swap is now simplified. Takes in big units (like SOL and BTC) instead of smol units (like satoshis and lamports). Accepts a slippage parameter.
+- Add a limit arg to placePerpOrder used when compute limit is reached
+
+### Changed 
+- Swap is now simplified. Takes in big units (like SOL and BTC) instead of smol units (like satoshis and lamports). Accepts a slippage parameter. 
+- PlacePerpOrder is now simplified. Takes in big units. No longer takes in maxQuoteQty.
 
 ## [0.1.0] - 2021-12-16