+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.2.0] - 2021-12-25 🎅
+
 ### Added
 - Withdraw & deposit now accept mint and deposit/withdraw directly to/from associated token account
 - Associated token account is created before withdrawal if such account does not exist